add swiftJSON, change the main view controller name
