+++ LAMO/controllers/MainViewController.swift
@@ -1,0 +1,43 @@
+//
+//  MainViewController.swift
+//  LAMO
+//
+//  Created by Nguyễn Lâm on 7/5/17.
+//  Copyright © 2017 Make School. All rights reserved.
+//
+
+import UIKit
+
+class MainViewController: UIViewController {
+    @IBOutlet weak var cameraButton: UIButton!
+    @IBOutlet weak var libraryButton: UIButton!
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+
+    override func didReceiveMemoryWarning() {
+        super.didReceiveMemoryWarning()
+        // Dispose of any resources that can be recreated.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destinationViewController.
+        // Pass the selected object to the new view controller.
+    }
+    */
+    @IBAction func cameraButtonTapped(_ sender: UIButton) {
+        print("camera button tapped")
+    }
+
+    @IBAction func libraryButtonTapped(_ sender: UIButton) {
+        print("library button tapped")
+    }
+}
