+++ LAMO/storyboards/Main.storyboard
@@ -9,10 +9,10 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <scenes>
-        <!--View Controller-->
+        <!--Photo API-->
         <scene sceneID="tne-QT-ifu">
             <objects>
-                <viewController id="BYZ-38-t0r" customClass="ViewController" customModule="LAMO" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController title="Photo API" id="BYZ-38-t0r" customClass="MainViewController" customModule="LAMO" customModuleProvider="target" sceneMemberID="viewController">
                     <layoutGuides>
                         <viewControllerLayoutGuide type="top" id="y3c-jy-aDJ"/>
                         <viewControllerLayoutGuide type="bottom" id="wfy-db-euE"/>
@@ -29,7 +29,7 @@
                                 </constraints>
                                 <state key="normal" backgroundImage="camera icon"/>
                                 <connections>
-                                    <action selector="cameraButtonTapped:" destination="BYZ-38-t0r" eventType="touchUpInside" id="wLB-Wm-L1J"/>
+                                    <action selector="cameraButtonTapped:" destination="BYZ-38-t0r" eventType="touchUpInside" id="x22-fF-cn7"/>
                                 </connections>
                             </button>
                             <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="IaH-xg-oUu">
@@ -40,7 +40,7 @@
                                 </constraints>
                                 <state key="normal" image="images icon"/>
                                 <connections>
-                                    <action selector="imagesButtonTapped:" destination="BYZ-38-t0r" eventType="touchUpInside" id="zbE-EB-wYq"/>
+                                    <action selector="libraryButtonTapped:" destination="BYZ-38-t0r" eventType="touchUpInside" id="Z9W-Ng-GzL"/>
                                 </connections>
                             </button>
                         </subviews>
@@ -53,8 +53,8 @@
                         </constraints>
                     </view>
                     <connections>
-                        <outlet property="cameraButton" destination="pgS-ao-NyW" id="aTj-lO-kyY"/>
-                        <outlet property="imagesButton" destination="IaH-xg-oUu" id="ogk-Fb-DI8"/>
+                        <outlet property="cameraButton" destination="pgS-ao-NyW" id="34j-h0-Fz1"/>
+                        <outlet property="libraryButton" destination="IaH-xg-oUu" id="xp4-19-7Xj"/>
                     </connections>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="dkx-z0-nzr" sceneMemberID="firstResponder"/>
